# reduce branch allocations